// src/Core/mipg/dto/create-mipg.dto.ts
export class CreateMipgDto {
  readonly user: number;
  readonly title: string;
  readonly category: string;
  readonly email: string;
  readonly url: string;
  logo?: string;
  readonly id?: string;
  readonly ip: string;
  readonly ip2?: string;
  readonly ip3?: string;
  readonly ip4?: string;
  readonly ip5?: string;
  ref: string;
  readonly karmozd: number;
  readonly authmobile: any;
  readonly authnationalcode: any;
  readonly authshahkar: any;
  readonly authnahab: any;
}